+++ src/codesmith/anvil/shell.clj
@@ -1,14 +1,28 @@
 (ns codesmith.anvil.shell
   (:require [clojure.string :as str]
-            [clojure.java.shell :as js]))
+            [clojure.java.shell :as js])
+  (:import (java.util List)))
 
-(defn sh [& args]
+(defn sh
+  "Execute the given shell command and return the output as a string."
+  [& args]
   (let [{:keys [exit out] :as result} (apply js/sh args)]
     (if (= exit 0)
       (str/trim out)
       (throw (ex-info "shell error"
                       (assoc result
                         :args args))))))
+
+(defn sh! [& args]
+  "Execute the given shell command and redirect the ouput/error to the standard output error; returns nil."
+  (let [^Process process (.. (ProcessBuilder. ^List args)
+                             (inheritIO)
+                             (start))
+        exit             (.waitFor process)]
+    (when (not= exit 0)
+      (throw (ex-info "shell error"
+                      {:exit 0
+                       :args args})))))
 
 (defn git-clean? []
   (= (count (sh "git" "status" "-s")) 0))
@@ -25,11 +39,11 @@
     0 7))
 
 (defn tag! [tag message]
-  (sh "git" "tag" "-am" message tag))
+  (sh! "git" "tag" "-am" message tag))
 
 (defn commit-all! [message]
-  (sh "git" "commit" "-am" message))
+  (sh! "git" "commit" "-am" message))
 
 (defn push-tags! []
-  (sh "git" "push")
-  (sh "git" "push" "--tags"))
+  (sh! "git" "push")
+  (sh! "git" "push" "--tags"))